# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is an API for managing a veterinary clinic. You can register cats, consultations and treatments.",
        "title": "Veterinarian API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1/vet",
    "paths": {
...
        "/visits": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all visits from the database, optionally filtered by vet, reason, or date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Get all visits with optional filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by veterinarian name",
                        "name": "vet",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by visit reason",
                        "name": "reason",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by date (format: YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.VisitHistoryResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve visits",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new visit entry in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Create a new visit",
                "parameters": [
                    {
                        "description": "Visit creation payload",
                        "name": "visit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.VisitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.VisitResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create visit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/visits/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a specific visit from the database by its ID, including associated treatments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Get visit by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Visit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.VisitHistoryResponse"
                        }
                    },
                    "404": {
                        "description": "Visit not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to find specific visit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing visit's information in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Update a visit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Visit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Visit update payload",
                        "name": "visit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.VisitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.VisitResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Visit not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update visit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a visit from the database by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visits"
                ],
                "summary": "Delete a visit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Visit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visit deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Visit not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete visit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.TreatmentHistoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "treatment_name": {
                    "type": "string"
                }
            }
        },
...
        "model.TreatmentResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "treatment_name": {
                    "type": "string"
                },
                "treatment_visit_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.VisitHistoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "visit_date": {
                    "type": "string"
                },
                "visit_reason": {
                    "type": "string"
                },
                "visit_treatments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TreatmentHistoryResponse"
                    }
                },
                "visit_vet": {
                    "type": "string"
                }
            }
        },
        "model.VisitRequest": {
            "type": "object",
            "properties": {
                "visit_cat_id": {
                    "type": "integer"
                },
                "visit_date": {
                    "type": "string"
                },
                "visit_reason": {
                    "type": "string"
                },
                "visit_vet": {
                    "type": "string"
                }
            }
        },
        "model.VisitResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "visit_cat_id": {
                    "type": "integer"
                },
                "visit_date": {
                    "type": "string"
                },
                "visit_reason": {
                    "type": "string"
                },
                "visit_treatments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TreatmentResponse"
                    }
                },
                "visit_vet": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}